simplify ViewModels

+++ app/src/main/java/i/am/shiro/amai/viewmodel/NhentaiViewModel.kt
@@ -1,5 +1,8 @@
 package i.am.shiro.amai.viewmodel
 
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.setValue
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.serialization.saved
@@ -7,15 +10,11 @@ import androidx.lifecycle.viewModelScope
 import androidx.room.withTransaction
 import i.am.shiro.amai.data.AmaiDatabase
 import i.am.shiro.amai.data.entity.CachedEntity
-import i.am.shiro.amai.data.view.CachedPreviewView
 import i.am.shiro.amai.network.GalleryDetailResponse
 import i.am.shiro.amai.network.Nhentai
 import i.am.shiro.amai.network.PaginatedResponse
 import i.am.shiro.amai.util.toEntity
-import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.launch
 import timber.log.Timber
@@ -34,59 +33,52 @@ class NhentaiViewModel(
 
     private var isComplete by handle.saved { false }
 
-    val books: StateFlow<List<CachedPreviewView>> = database.cachedPreviewDao
-        .getAll()
+    val books = database.cachedPreviewDao.getAll()
         .stateIn(viewModelScope, SharingStarted.WhileSubscribed(5000), emptyList())
 
-    private val _isLoading = MutableStateFlow(false)
-    val isLoading = _isLoading.asStateFlow()
+    var isLoading by mutableStateOf(false)
+        private set
 
     init {
         if (page == 0) {
-            onRefresh()
-        }
-    }
-
-    fun onScrollToBottom() {
-        if (isComplete || _isLoading.value) return
-        fetchRemotePage()
-    }
-
-    fun onRefresh() {
-        page = 0
-        isComplete = false
-
-        viewModelScope.launch {
-            deleteLocal()
             fetchRemotePage()
         }
     }
 
-    fun onSort(sort: Nhentai.Sort) {
+    fun loadMore() {
+        if (isComplete || isLoading) return
+        fetchRemotePage()
+    }
+
+    fun refresh() {
+        page = 0
+        isComplete = false
+        fetchRemotePage()
+    }
+
+    fun sort(sort: Nhentai.Sort) {
         if (this.sort == sort) return
         this.sort = sort
-        onRefresh()
+        refresh()
     }
 
-    fun onSearch(query: String) {
+    fun search(query: String) {
         if (this.query == query) return
         this.query = query
-        onRefresh()
+        refresh()
     }
 
-    private suspend fun deleteLocal() {
-        try {
-            database.cachedDao.deleteAll()
-            database.bookDao.deleteOrphan()
-        } catch (e: Exception) {
-            Timber.e(e)
-        }
-    }
-
+    // FIXME sort is ignored on empty query
     private fun fetchRemotePage() {
         viewModelScope.launch {
-            _isLoading.value = true
+            isLoading = true
             try {
+                // delete cache first before loading page 1
+                if (page == 0) {
+                    database.cachedDao.deleteAll()
+                    database.bookDao.deleteOrphan()
+                }
+
                 if (query.isEmpty()) {
                     val response = nhentaiApi.getAll(page + 1)
                     onSearchSuccess(response)
@@ -101,7 +93,7 @@ class NhentaiViewModel(
             } catch (e: Exception) {
                 Timber.e(e)
             } finally {
-                _isLoading.value = false
+                isLoading = false
             }
         }
     }
